# 🏗 Cache `node_modules` during CI to speed up installs