--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -67,9 +67,19 @@ commands:
       - run:
           name: 'Configure Hosts'
           command: cat ./build-system/test-configs/hosts | sudo tee -a /etc/hosts
+      - restore_cache:
+          name: 'Restore Node Modules'
+          keys:
+            - node-modules-cache-{{ checksum "package-lock.json" }}
+            - node-modules-cache-
       - run:
           name: 'Install Dependencies'
           command: ./.circleci/install_dependencies.sh
+      - save_cache:
+          name: 'Save Node Modules'
+          key: node-modules-cache-{{ checksum "package-lock.json" }}
+          paths:
+            - node_modules
   install_chrome:
     steps:
       - browser-tools/install-chrome: