--- a/.github/workflows/cross-browser-tests.yml
+++ b/.github/workflows/cross-browser-tests.yml
@@ -23,7 +23,9 @@ jobs:
         uses: actions/cache@v2
         with:
           path: node_modules
-          key: node_modules-${{ matrix.platform }}-${{ hashFiles('package-lock.json') }}
+          key: node-modules-cache-${{ matrix.platform }}-${{ hashFiles('package-lock.json') }}
+          restore-keys: |
+            node-modules-cache-${{ matrix.platform }}-
       - name: Install Dependencies
         run: bash ./.github/workflows/install_dependencies.sh
       - name: Build and Test