--- a/.github/workflows/install_dependencies.sh
+++ b/.github/workflows/install_dependencies.sh
@@ -33,6 +33,6 @@ echo $(GREEN "Installing gulp-cli...")
 npm install --global gulp-cli
 
 echo $(GREEN "Installing dependencies...")
-npm ci
+npm install
 
 echo $(GREEN "Successfully installed all project dependencies.")