**Background:**

Package installation during CI has gotten slower over time, especially on Windows and macOS. Reducing the number of dependencies will need careful rewriting of code. Meanwhile, we can experiment with caching.

#32429 tried to persist the NPM cache directory (typically localted at `~/.npm`) so that `npm ci` (a.k.a. `clean-install`) could use it to freshly recreate `node_modules`. Speeds either remained stagnant or [became worse](https://github.com/ampproject/amphtml/pull/32429#issuecomment-773583411).

**PR Highlights:**

- Instead of `~/.npm`, we cache `node_modules`
- `npm install` will overwrite only what has changed in `node_modules`
- Primary key is a known prefix + a hash of `package-lock.json` (exact match => all cached packages are fresh)
- Secondary key is just the known prefix (slightly older cache => `npm install` will refresh what has changed)

**Results:**

| **VM / Install time** | **CircleCI Linux (L)** | **GH Actions Linux** | **GH Actions macOS** | **GH Actions Windows** |
| --- | ---  |--- | --- |--- |
| **Cold cache** | [1m 41s](https://app.circleci.com/pipelines/github/ampproject/amphtml/1292/workflows/34980c82-70a3-4bed-999d-fd3d36c88e93/jobs/11977) | [1m 26s](https://github.com/ampproject/amphtml/runs/1834510931?check_suite_focus=true) | [3m 26s](https://github.com/ampproject/amphtml/runs/1834510950?check_suite_focus=true) | [4m 34s](https://github.com/ampproject/amphtml/runs/1834510973?check_suite_focus=true) |
| **Warm cache** | [41s](https://app.circleci.com/pipelines/github/ampproject/amphtml/1295/workflows/cf57693f-906e-4c29-bb1e-f0df0a99ad6b/jobs/12034) | [43s](https://github.com/ampproject/amphtml/runs/1834698035?check_suite_focus=true) | [38s](https://github.com/ampproject/amphtml/runs/1834698050?check_suite_focus=true) | [4m 15s](https://github.com/ampproject/amphtml/runs/1834698068?check_suite_focus=true) |

**Update:** Abandoned because of https://github.com/ampproject/amphtml/pull/32439#issuecomment-773653012.