@rsimha 2021-02-04
> Meanwhile, can y'all think of any downsides to this approach?

A few seconds after I sent this out, there's [this familiar `@babel` error](https://github.com/ampproject/amphtml/runs/1834698068?check_suite_focus=true#step:5:68), most likely because we're using `npm install` instead of `npm ci`.

I guess I'm going to have to abandon this PR as well :/